@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 class BackgroundImage extends StatelessWidget {
 
   const BackgroundImage({
-    Key key
+    Key? key
   }) : super(key: key);
 
   final _imagePath = 'assets/img/bg_mertola.jpeg';
