@@ -2,13 +2,12 @@ import 'package:app_ppm/components/BackgroundImage.dart';
 import 'package:app_ppm/providers/firebaseProvider.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-
-
+import 'package:app_ppm/utils/routes.dart' as my_routes;
 
 
 
 class Home extends StatelessWidget {
-  Home({Key key}) : super(key: key);
+  Home({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -40,7 +39,7 @@ class Home extends StatelessWidget {
                             children: [
                               Column(
                                 children: [
-                                  Text("${Provider.of<firebaseProvider>(context).xpto2((index + 1).toString())}" + " . " + "${Provider.of<firebaseProvider>(context).xpto()}",
+                                  Text("${Provider.of<firebaseProvider>(context, listen: false).xpto2((index + 1).toString())}" + " . " + "${Provider.of<firebaseProvider>(context).xpto()}",
                                     style: TextStyle(
                                       fontFamily: 'Ubuntu',
                                       fontSize: 25,
