@@ -1,3 +1,4 @@
-const SCREEN_HOME = '/home';
+const SCREEN_HOME   = '/home';
+const SCREEN_SPLASH = '/splash_screen';
 
 
